Ajout des commentaires et mise au propre du code

+++ back/controllers/user.js
@@ -3,27 +3,40 @@ const fs = require("fs");
 const User = require("../models/User.js");
 const bcrypt = require("bcrypt");
 
+//Ajoute une modifie l'image de profil d'origine
 exports.avatar = (req, res, next) => {
   const userID = req.params["id"];
-  const avatar = `${req.protocol}://${req.get("host")}/images/${req.file.filename}`;
+  const avatar = `${req.protocol}://${req.get("host")}/images/${
+    req.file.filename
+  }`;
 
-  database.query(`SELECT pictureName FROM crud.users WHERE idusers=${userID}`)
-    .then((result) => { if (result[0][0].pictureName != null) {const filename = result[0][0].pictureName.split("/images/")[1];
-    console.log(filename);
-    
-     fs.unlink(`images/${filename}`, () => {
-       database.query(`UPDATE crud.users SET pictureName='${avatar}' WHERE idusers=${userID}`)
-    .then(() => res.status(200).json(avatar))
-    .catch((error) => res.status(400).json({error}))
-     })}else{
-      database.query(`UPDATE crud.users SET pictureName='${avatar}' WHERE idusers=${userID}`)
-      .then(() => res.status(200).json(avatar))
-      .catch((error) => res.status(400).json({error}))
-     }})
-    .catch((error) => res.status(500).json({error}));
+  database
+    .query(`SELECT pictureName FROM crud.users WHERE idusers=${userID}`)
+    .then((result) => {
+      if (result[0][0].pictureName != null) {
+        const filename = result[0][0].pictureName.split("/images/")[1];
 
+        fs.unlink(`images/${filename}`, () => {
+          database
+            .query(
+              `UPDATE crud.users SET pictureName='${avatar}' WHERE idusers=${userID}`
+            )
+            .then(() => res.status(200).json(avatar))
+            .catch((error) => res.status(400).json({ error }));
+        });
+      } else {
+        database
+          .query(
+            `UPDATE crud.users SET pictureName='${avatar}' WHERE idusers=${userID}`
+          )
+          .then(() => res.status(200).json(avatar))
+          .catch((error) => res.status(400).json({ error }));
+      }
+    })
+    .catch((error) => res.status(500).json({ error }));
 };
 
+//Récupère les informations de l'utilisateur à la connection
 exports.getInfo = (req, res, next) => {
   database
     .query(
@@ -45,70 +58,85 @@ exports.getInfo = (req, res, next) => {
       }
     });
 };
-
+//Modifie les information de l'utilisateur
 exports.updateInfo = (req, res, next) => {
-    let userID = req.params["id"];
-    if (req.body.password.length > 6) {
-        bcrypt
-          .hash(req.body.password, 10)
-          .then((hash) => {
-            let user = new User(
-              req.body.firstname,
-              req.body.lastname,
-              req.body.email,
-              req.body.picturename,
-              hash
-            ); console.log(userID);
-            database
-              .query(`UPDATE crud.users SET firstname="${req.body.firstname}",
+  let userID = req.params["id"];
+  if (req.body.password.length > 6) {
+    bcrypt
+      .hash(req.body.password, 10)
+      .then((hash) => {
+        let user = new User(
+          req.body.firstname,
+          req.body.lastname,
+          req.body.email,
+          req.body.picturename,
+          hash
+        );
+        console.log(userID);
+        database
+          .query(
+            `UPDATE crud.users SET firstname="${req.body.firstname}",
                                             lastname="${req.body.lastname}",
                                             email="${req.body.email}",
                                             pass="${hash}"
-                                            WHERE idusers="${userID}"`)
-              .then((result) => {
-                return res.status(201).json({ message: "Utilisateur créé" });
-              })
-              .catch((error) => res.status(500).json({ error }));
+                                            WHERE idusers="${userID}"`
+          )
+          .then(() => {
+            return res.status(201).json({ message: "Utilisateur créé" });
           })
           .catch((error) => res.status(500).json({ error }));
-      } else {
-        return res.status(403).json({
-          message: "Le mot de passe doit contenir au minimum 6 caractères",
-        });
-      }
-        
+      })
+      .catch((error) => res.status(500).json({ error }));
+  } else {
+    return res.status(403).json({
+      message: "Le mot de passe doit contenir au minimum 6 caractères",
+    });
+  }
 };
-
+//supprime toutes les informations de l'utilisateur
 exports.deleteProfile = (req, res, next) => {
-    let userID = req.params["id"]; 
-    
-    database.query(`SELECT pictureName FROM crud.users WHERE idusers=${userID}`)
-    .then((result) => { if (result[0][0].pictureName != null) {const filename = result[0][0].pictureName.split("/images/")[1];
-    console.log(filename);
-    
-     fs.unlink(`images/${filename}`, () => {
-       database.query(`DELETE FROM crud.users WHERE idusers=${userID}`)
-    .then(() => res.status(200).json({message: "utilisateur supprimé"}))
-    .catch((error) => res.status(400).json({error}))
-     })}else{
-      database.query(`DELETE FROM crud.users WHERE idusers=${userID}`)
-      .then(() => res.status(200).json({message: "utilisateur supprimé"}))
-      .catch((error) => res.status(400).json({error}))
-     }})
-    .catch((error) => res.status(500).json({error}));
+  let userID = req.params["id"];
 
-    
+  database
+    .query(`SELECT pictureName FROM crud.users WHERE idusers=${userID}`)
+    .then((result) => {
+      if (result[0][0].pictureName != null) {
+        const filename = result[0][0].pictureName.split("/images/")[1];
+        console.log(filename);
+
+        fs.unlink(`images/${filename}`, () => {
+          database
+            .query(`DELETE FROM crud.users WHERE idusers=${userID}`)
+            .then(() =>
+              res.status(200).json({ message: "utilisateur supprimé" })
+            )
+            .catch((error) => res.status(400).json({ error }));
+        });
+      } else {
+        database
+          .query(`DELETE FROM crud.users WHERE idusers=${userID}`)
+          .then(() => res.status(200).json({ message: "utilisateur supprimé" }))
+          .catch((error) => res.status(400).json({ error }));
+      }
+    })
+    .catch((error) => res.status(500).json({ error }));
 };
+
+//Récupère les informations utilisateur pour le profil
 exports.getUserInfo = (req, res, next) => {
   let userID = req.params["id"];
-  database.query(`SELECT firstname, lastname, pictureName FROM crud.users WHERE idusers=${userID}`)
-  .then((result) => res.status(200).json(result[0][0].pictureName))
-  .catch((error) => res.status(403).json({error}))
-}
+  database
+    .query(
+      `SELECT firstname, lastname, pictureName FROM crud.users WHERE idusers=${userID}`
+    )
+    .then((result) => res.status(200).json(result[0][0].pictureName))
+    .catch((error) => res.status(403).json({ error }));
+}; 
 
+//Récupère tout les utilisateurs
 exports.getAllUsers = (req, res, next) => {
-    database.query(`SELECT * FROM crud.users`)
-    .then(() => res.status(200).json({message: "liste récupérée"}))
-    .catch((error) => res.status(404).json(error))
-  
-}
+  database
+    .query(`SELECT * FROM crud.users`)
+    .then(() => res.status(200).json({ message: "liste récupérée" }))
+    .catch((error) => res.status(404).json(error));
+};
